requirements.txt
appdirs==1.4.4
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.3
cycler==0.11.0
docker-pycreds==0.4.0
einops==0.7.0
fenics-dijitso==2019.1.0
fenics-dolfin==2019.1.0
fenics-ffc==2019.1.0
fenics-fiat==2019.1.0
fenics-ufl==2019.1.0
filelock==3.9.0
fonttools==4.43.0
gitdb==4.0.10
GitPython==3.1.37
gmpy2==2.1.2
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lnpde==0.0.1
MarkupSafe==2.1.1
matplotlib==3.8.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpi4py==3.1.4
mpmath==1.3.0
munkres==1.1.4
networkx==3.1
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pathtools==0.1.2
patsy==0.5.3
petsc4py==3.17.1
Pillow==9.4.0
pip==23.2.1
pkgconfig==1.5.5
ply==3.11
protobuf==4.24.3
psutil==5.9.5
pybind11==2.9.2
pybind11-global==2.9.2
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.1
seaborn==0.12.2
sentry-sdk==1.31.0
setproctitle==1.3.2
setuptools==68.0.0
sip==6.7.11
six==1.16.0
slepc4py==3.17.1
smmap==5.0.1
statsmodels==0.14.0
sympy==1.11.1
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchdiffeq==0.2.3
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
unicodedata2==15.1.0
urllib3==1.26.16
wandb==0.15.11
wheel==0.41.2
